Add testing of finish_course.php

+++ finish_course.php
@@ -14,6 +14,11 @@ if (($event == "") || ($competitor_id == "")) {
 }
 
 $competitor_path = "./" . $event . "/Competitors/" . $competitor_id;
+
+if (!file_exists($competitor_path) || !file_exists("./{$event}/Courses/{$course}/controls.txt")) {
+  error_and_exit("<p>ERROR: Event \"{$event}\" or competitor \"{$competitor}\" appears to be no longer appears valid, please re-register and try again.\n");
+}
+
 $control_list = file("./${event}/Courses/${course}/controls.txt");
 $control_list = array_map('trim', $control_list);
 //echo "Controls on the ${course} course.<br>\n";
@@ -24,32 +29,39 @@ if (!file_exists("${competitor_path}/start")) {
   error_and_exit("<p>Course " . ltrim($course, "0..9-") . " not yet started.\n<br>Please scan the start QR code to start a course.\n");
 }
 
-// See how many controls have been completed
-$controls_done = scandir("./${competitor_path}");
-$controls_done = array_diff($controls_done, array(".", "..", "course", "name", "next", "start", "finish", "extra", "dnf")); // Remove the annoying . and .. entries
-// echo "<br>Controls done on the ${course} course.<br>\n";
-// print_r($controls_done);
-
-// Are we at the right control?
-$number_controls_found = count($controls_done);
-$number_controls_on_course = count($control_list);
-// echo "<br>At control ${control_id}, expecting to be at " . $control_list[$number_controls_found] . "--\n";
-if ($number_controls_found != $number_controls_on_course) {
-    $error_string .= "<p>Not all controls found, found ${number_controls_found} controls, expected ${number_controls_on_course} controls.\n";
-    file_put_contents($competitor_path . "/dnf", $error_string, FILE_APPEND);
+if (!file_exists("{$competitor_path}/finish")) {
+  // See how many controls have been completed
+  $controls_done = scandir("./${competitor_path}");
+  $controls_done = array_diff($controls_done, array(".", "..", "course", "name", "next", "start", "finish", "extra", "dnf")); // Remove the annoying . and .. entries
+  // echo "<br>Controls done on the ${course} course.<br>\n";
+  // print_r($controls_done);
+  
+  // Are we at the right control?
+  $number_controls_found = count($controls_done);
+  $number_controls_on_course = count($control_list);
+  // echo "<br>At control ${control_id}, expecting to be at " . $control_list[$number_controls_found] . "--\n";
+  if ($number_controls_found != $number_controls_on_course) {
+      $error_string .= "<p>Not all controls found, found ${number_controls_found} controls, expected ${number_controls_on_course} controls.\n";
+      file_put_contents($competitor_path . "/dnf", $error_string, FILE_APPEND);
+  }
+  
+  $now = time();
+  file_put_contents($competitor_path . "/finish", strval($now));
+  $course_started_at = file_get_contents($competitor_path . "/start");
+  $time_taken = $now - $course_started_at;
+  if (!file_exists("./${event}/Results/${course}")) {
+    mkdir("./${event}/Results/${course}");
+  }
+  $result_filename = sprintf("%06d,%s", $time_taken, $competitor_id);
+  file_put_contents("./${event}/Results/${course}/${result_filename}", "");
+}
+else {
+  $error_string .= "<p>Second scan of finish?  Finish time not updated.\n";
+  $course_started_at = file_get_contents($competitor_path . "/start");
+  $course_finished_at = file_get_contents($competitor_path . "/finish");
+  $time_taken = $course_finished_at - $course_started_at;
 }
 
-$now = time();
-file_put_contents($competitor_path . "/finish", strval($now));
-$course_started_at = file_get_contents($competitor_path . "/start");
-$time_taken = $now - $course_started_at;
-if (!file_exists("./${event}/Results/${course}")) {
-  mkdir("./${event}/Results/${course}");
-}
-$result_filename = sprintf("%06d,%s", $time_taken, $competitor_id);
-file_put_contents("./${event}/Results/${course}/${result_filename}", "");
-
-// Don't forget to update the results file
 
 // Clear the cookies, ready for another course registration
 setcookie("competitor_id", $competitor_id, 1);
